v1.1 update.
An update after 3 years since first created. On usages and the looking,
the app doesn't seem very different. Features are nearly just the same,
but with a big structure change on the code side using MVVM pattern.
A small improvement on feature is, the cursor will keep following first
tracked (lifting-forward) hand regardless of another hand lifting
afterwards. Previously, it was set to only follow right hand when both
hand are tracked.