Modifications to be ergonomically usable in seated position for gaming, implement left,middle,right mouse button
Deadzone is configurable to be ratio of forearm distance (elbow to wrist) along the Z-axis from the spine shoulder joint which is visible in a seated position where the spine base joint may be obscured by a desk. The default value allows for placement of joystick, throttle, keyboard and mouse in the deadzone.

Define left, middle, and right mouse buttons as grab with the thumb position relative to the hand horizontally, diagonally, and vertically.

Use the wrist position relative to the respective shoulder to control cursor movement defined by a rectangle based on forearm length for ergonomics. This reduces mouse position jitter when grabbing which affects hand position.

Directly forward from the shoulder corresponds to the center of the bottom of the screen.  The top of the screen corresponds to one forearm length up. The sides of the screen corresponds to one forearm length left and right. Using these bounds reduces the amount of movement required vs using the spinebase as the center. These rectangular coordinates are also visible in a seated position.

A universal stop gesture (crossing the forearms at the wrists with open hands) is implemented to disable the mouse tracking